food-backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
djoser
